+++ admin/src/main/resources/io/buoyant/admin/js/src/router_server.js
@@ -19,6 +19,7 @@ define([
         {
           description: "Pending",
           metricSuffix: "load",
+          isGauge: true,
           query: Query.serverQuery().withRouter(routerName).withServer(serverName).withMetric("load").build()
         },
         {
@@ -65,7 +66,7 @@ define([
         var serverData = Query.filter(defn.query, data);
 
         if (!_.isEmpty(serverData)) {
-          defn.value = serverData[0].delta;
+          defn.value = defn.isGauge ? serverData[0].value : serverData[0].delta;
           lookup[defn.metricSuffix] = defn.value;
         }
 
@@ -96,9 +97,7 @@ define([
       var chart = SuccessRateGraph($chartEl, "#4AD8AC");
 
       var metricsHandler = function(data) {
-        var filteredData = _.filter(data.specific, function (d) { return d.name.indexOf(routerName) !== -1 });
-        var transformedData = processData(filteredData, routerName, server.label);
-
+        var transformedData = processData(data.specific, routerName, server.label);
         renderServer($metricsEl, server, transformedData);
         chart.updateMetrics(getSuccessRate(transformedData));
       }