+++ linkerd/protocol/http/src/integration/scala/io/buoyant/linkerd/protocol/TlsNoValidationTest.scala
@@ -21,7 +21,7 @@ class TlsNoValidationTest extends FunSuite with Awaits {
              |- protocol: http
              |  dtab: |
              |    /p/dog => /$$/inet/127.1/${dog.port} ;
-             |    /http/1.1/GET/clifford => /p/dog ;
+             |    /svc/clifford => /p/dog ;
              |  servers:
              |  - port: 0
              |  client: